// types.d.ts
export interface BlogPost {
    id: string;
    title: string;
    excerpt: string;
    slug: string;
    coverImage?: CoverImage;
    publishedAt: string;
    readTimeInMinutes: number;
    author: Author;
    tags?: Tag[];
}
/** Detailed blog post with content */
export interface BlogPostDetail extends BlogPost {
    content?: PostContent;
}
/** Cover image data */
export interface CoverImage {
    url?: string;
}
/** Author information */
export interface Author {
    name: string;
    username: string;
    profilePicture?: string;
}
/** Tag data */
export interface Tag {
    name: string;
    slug?: string;
}